requirements.txt
requests~=2.24.0
python-dotenv==0.14.0
python-json-logger==0.1.11
watchdog==0.10.3
pika==0.13.1
beautifulsoup4==4.8.0
bs4==0.0.1
